+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* Minor updates to config comments
+
 ## 3.3.1
 * Added a config option to make mobs not affected by (most of) Enhanced AI features based off their spawn type
   * This adds a new NBT `enhancedai:unaffected_by_features` that if true will prevent the mob from getting any changes from Enhanced AI features